pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Fabio Bignoto Palacio <fabiopalacio14@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
django-rest-framework = "^0.1.0"
pytest = "^8.3.2"
pytest-django = "^4.8.0"
django-extensions = "^3.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
factory-boy = "^3.3.0"
coverage = "^7.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
